backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
faiss-cpu
numpy
python-dotenv
google-genai
pydantic
requests